Change interface of DAO.find();
Not finding an object in a DAO is no longer an error.  The promised
returned from DAO.find() resolves with null when the object isn't found
rather than rejecting with an exception.

The promise is only rejected in actual error situations like internal
DAO exceptions, or invalid arguments.

+++ src/foam/u2/view/TreeView.js
@@ -131,6 +131,11 @@ foam.CLASS({
       var self = this;
       var dao  = this.__context__[this.relationship.targetDAOKey];
       dao.find(id).then(function(obj) {
+        if ( ! obj ) return null;
+
+        // TODO: We shouldn't have to remove then put,
+        // We currently have to because the FLOW editor is not updating properly
+        // on a put event for an object that it already has.
         dao.remove(obj).then(function() {
           self.data[self.relationship.forwardName].put(obj).then(function(obj) {
             self.onObjDrop(obj, id);
